fix weight to float32

+++ ochat/models/unpadded_gemma2.py
@@ -77,7 +77,7 @@ def rms_norm(
 
     variance = hidden_states.pow(2).mean(-1, keepdim=True)
     hidden_states = hidden_states * torch.rsqrt(variance + variance_epsilon)
-    return (1 + weight) * hidden_states.to(input_dtype)
+    return (1 + weight.to(torch.float32)) * hidden_states.to(input_dtype)
 
 
 def rotate_half(x: torch.Tensor):
